backend/requirements.txt
absl-py==0.8.1
astor==0.8.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
Django==2.2.6
django-mysql==3.3.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gast==0.2.2
google-auth==1.8.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
implicit==0.4.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mysqlclient==1.4.6
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-hub==0.7.0
termcolor==1.1.0
tqdm==4.40.2
urllib3==1.25.7
uWSGI==2.0.18
Werkzeug==0.16.0
wrapt==1.11.2
coveralls==1.8.2
